# app/contracts/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Zhixue Mate Agent API",
    "version": "api-contract-v0.3",
    "description": "知学 Mate 前后端唯一接口契约。字段变更必须先升契约版本并通知前端。 声明层已于 2026-09-22 按实跑结果补全：补齐 5 个已实现端点、所有 operation 的 405、register 的 409、exercises 的 4 个查询参数与 400、knowledge-bases 的 400、可空字段的 nullable、版本头参数，并移除死枚举值 LLM_FALLBACK。实现未改动。 KnowledgeBaseDetailResponse.documents 已于 2026-09-22 补声明（后端返回、前端在读，此前契约未声明）。 LegacyChatResponse.card.summary 已于 2026-09-22 补声明（前端在读、后端已补返回，此前两边都缺）。"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "本地开发"
    },
    {
      "url": "http://{host}:5000",
      "description": "局域网联调",
      "variables": {
        "host": {
          "default": "192.168.1.100"
        }
      }
    }
  ],
  "security": [],
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "summary": "注册账号（昵称 + 年级，无需密码）",
        "description": "注册即下发 userId 与 token，并为新账号预置一份空画像与起始计划。演示身份 demo-user 无需注册。",
        "operationId": "authRegister",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "该手机号已注册（errorCode=CONFLICT）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          }
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "用已保存的 userId + token 恢复登录态",
        "operationId": "authLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          }
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "summary": "退出登录（仅吊销当前 token）",
        "operationId": "authLogout",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "已退出",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "summary": "读取当前登录用户（前端启动时校验本地 token 是否仍有效）",
        "operationId": "authMe",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "当前用户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUserResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          }
        ]
      }
    },
    "/api/v1/auth/account": {
      "delete": {
        "summary": "注销账号（合规要求：标记停用并清空全部会话）",
        "operationId": "authDeleteAccount",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "账号已注销",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          }
        ]
      }
    },
    "/api/v1/auth/login-or-register": {
      "post": {
        "summary": "登录优先，注册兜底",
        "description": "昵称已存在则登录该账号（不新建），不存在才建号。登录页主按钮应调此接口，避免同一昵称重复建号。",
        "operationId": "loginOrRegister",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录了已有账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginOrRegisterResponse"
                }
              }
            }
          },
          "201": {
            "description": "新建了账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginOrRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          }
        ]
      }
    },
...
    "/api/v1/auth/login-with-huawei": {
      "post": {
        "summary": "华为账号一键登录（登录优先，注册兜底）",
        "description": "客户端用 @kit.AccountKit 取得 OpenID 后调用本接口。首次登录自动建号，用户无需填昵称/手机号/验证码。⚠️ 本接口**不校验 OpenID 真伪**（需服务端调华为接口验签），当前定位为演示可用、生产需补验签，见 docs/13。",
        "operationId": "loginWithHuawei",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthHuaweiLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录了已有账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginOrRegisterResponse"
                }
              }
            }
          },
          "201": {
            "description": "首次登录，已自动建号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginOrRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "缺少 openId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          }
        ]
      }
    },
    "/api/v1/auth/send-code": {
      "post": {
        "summary": "发送手机号验证码（开发模式回显）",
        "description": "同号 60 秒内限发一次，滚动 24 小时最多 10 次。未配置短信服务商时仅在非生产环境回显 devCode，smsDelivered=false，绝不伪装已发送。",
        "operationId": "authSendCode",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneCodeSendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "验证码已签发",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneCodeSendResponse"
                }
              }
            }
          },
          "400": {
            "description": "手机号格式错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "触发重发或每日限流，或手机号处于锁定状态",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "短信服务未接入或生产环境未配置",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          }
        ]
      }
    },
    "/api/v1/auth/verify-code": {
      "post": {
        "summary": "校验验证码并登录或自动注册",
        "description": "验证码成功后立即失效；手机号已存在则登录，不存在则创建账号并预置起始画像。",
        "operationId": "authVerifyCode",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneCodeVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录已有账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginOrRegisterResponse"
                }
              }
            }
          },
          "201": {
            "description": "验证码通过并新建账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginOrRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "验证码错误、过期、已使用或参数非法",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "手机号因连续错误被锁定",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          }
        ]
      }
    },
...
    "/api/v1/auth/change-password": {
      "post": {
        "summary": "修改 / 首次设置密码",
        "description": "需 Bearer token。账号已设密码时必须提供正确的 oldPassword；从未设过密码（验证码建号）时为首次设置，无需旧密码。本接口不吊销其它会话。",
        "operationId": "authChangePassword",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已修改 / 已设置",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthChangePasswordResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误（新密码长度等）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "405": {
            "description": "请求方法不被该路径支持（后端统一返回 JSON 错误体，而非 HTML）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/ContractVersionHeader"
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "可选鉴权。/api/v1/** 不带 token 时按演示/游客身份（demo-user）处理，以保证演示主链不受影响。"
      }
    },
    "parameters": {
...
      "ContractVersionHeader": {
        "name": "X-API-Contract-Version",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "接口契约版本闸门。**可选**：不带头时后端放行；带上但值不等于当前版本时返回 409 CONTRACT_VERSION_MISMATCH。当前版本见 info.version。"
      }
    },
    "schemas": {
...
      "AuthRegisterRequest": {
        "type": "object",
        "required": [
          "nickname",
          "phone"
        ],
        "properties": {
          "nickname": {
            "type": "string",
            "minLength": 1,
            "maxLength": 16,
            "description": "昵称，去除首尾空白后不能为空"
          },
          "grade": {
            "type": "string",
            "nullable": true,
            "enum": [
              "大一",
              "大二",
              "大三",
              "大四",
              "大五",
              "研一",
              "研二",
              "研三",
              "博士"
            ],
            "description": "年级，可选"
          },
          "phone": {
            "type": "string",
            "description": "手机号。注册必填；生产登录应先调用 send-code + verify-code。当前 /auth/login 的免验证码手机号登录保留给演示与单机环境。"
          },
          "password": {
            "type": "string",
            "nullable": true,
            "minLength": 6,
            "maxLength": 72,
            "description": "可选。设置后该账号可用密码登录；服务端只保存 Werkzeug 生成的加盐哈希，明文不落库也不回传。未设置（例如走验证码建号）时，该账号的密码登录会被如实拒绝并提示改用验证码登录 —— 不会接受任意密码。"
          }
        }
      },
      "AuthLoginRequest": {
        "type": "object",
        "required": [],
        "properties": {
          "userId": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "nickname": {
            "type": "string",
            "nullable": true,
            "description": "按昵称登录。与 userId/token 二选一；两者都提供时优先按凭据恢复。账号设过密码时需同时提供 password。昵称本身不是秘密，此形式仅适用于演示/单机场景。"
          },
          "phone": {
            "type": "string",
            "description": "按手机号登录。与 userId/token、nickname 三选一；分派优先级：凭据 > 手机号 > 昵称。"
          },
          "password": {
            "type": "string",
            "nullable": true,
            "description": "配合 phone / nickname 使用。账号设过密码时**必须**正确，否则返回 401 UNAUTHORIZED；从未设过密码的账号会被拒绝并提示改用「验证码登录」。"
          }
        }
      },
...
      "AuthSessionResponse": {
        "type": "object",
        "required": [
          "user",
          "token",
          "expiresAt"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          },
          "token": {
            "type": "string",
            "description": "长期会话票据（默认 30 天）。存设备 preferences，请求时放入 Authorization: Bearer"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuthUserResponse": {
        "type": "object",
        "required": [
          "user"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        }
      },
      "AuthStatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ok",
              "deactivated"
            ]
          },
          "userId": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "errorCode",
          "message"
        ],
        "properties": {
          "errorCode": {
            "type": "string",
            "enum": [
              "BAD_REQUEST",
              "NOT_FOUND",
              "CONFLICT",
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "CONTRACT_VERSION_MISMATCH",
              "INTERNAL_ERROR",
              "RATE_LIMITED",
              "DAILY_LIMIT_REACHED",
              "PHONE_LOCKED",
              "INVALID_CODE",
              "CODE_EXPIRED",
              "CODE_NOT_FOUND",
              "SMS_NOT_CONFIGURED",
              "SMS_PROVIDER_UNAVAILABLE"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "nullable": true
          }
        }
      },
...
      "AuthLoginOrRegisterResponse": {
        "type": "object",
        "required": [
          "user",
          "token",
          "expiresAt",
          "created"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          },
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "created": {
            "type": "boolean",
            "description": "true=本次新建账号；false=登录了已有账号"
          }
        }
      },
...
      "AuthHuaweiLoginRequest": {
        "type": "object",
        "required": [
          "openId"
        ],
        "properties": {
          "openId": {
            "type": "string",
            "description": "华为账号 OpenID（应用内唯一）。由客户端 @kit.AccountKit 的 LoginWithHuaweiIDRequest 取得。"
          },
          "unionId": {
            "type": "string",
            "nullable": true,
            "description": "华为账号 UnionID（开发者账号内唯一），保留用于多应用打通"
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "grade": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PhoneCodeSendRequest": {
        "type": "object",
        "required": [
          "phone"
        ],
        "properties": {
          "phone": {
            "type": "string",
            "description": "11 位中国大陆手机号，支持 +86 与分隔符写法"
          }
        }
      },
      "PhoneCodeSendResponse": {
        "type": "object",
        "required": [
          "phoneMasked",
          "expiresInSeconds",
          "resendAfterSeconds",
          "smsDelivered"
        ],
        "properties": {
          "phoneMasked": {
            "type": "string",
            "example": "138****0000"
          },
          "expiresInSeconds": {
            "type": "integer",
            "example": 300
          },
          "resendAfterSeconds": {
            "type": "integer",
            "example": 60
          },
          "smsDelivered": {
            "type": "boolean",
            "description": "true 才表示短信服务商已实际发送"
          },
          "devCode": {
            "type": "string",
            "nullable": true,
            "description": "仅开发模式回显；生产模式绝不返回"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "未登录或登录已失效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  }
}